termux-tools: handle mirror groups instead of individual mirrors
Group mirrors per geographical location and make users select which
group to use with termux-change-repo.  If no group have been chosen
with termux-change-repo, then all are used.

The environmental variable TERMUX_PKG_NO_MIRROR_SELECT can still be
used to lock a specific mirror.

The weight system is removed with this change, each mirror in a group
is given the same weight.

The version scheme is also changed from 0.XXX to 1.X, due to me not
finding it intuitive if for example 0.101 is higher or lower than
0.99.